fix(UIPointer): UI not responding after application loses focus
In Unity 5.5 and above the default `EventSystem` pauses its processing
whenever `OnApplicationFocus(false)` is called on it. The solution is
to subclass the default `EventSystem` and *not* call the base
implementation in `OnApplicationFocus`. The rest of this code works
around some problems when changing the `EventSystem` at runtime while
another one is active and referenced by other objects. See the code
comments for more info on this.