fix(Interaction): ensure trigger collision check is done every frame
Previously, the `triggerIsColliding` bool was only set if an object
wasn't being touched which would cause the `StopTouching` method to
execute which caused problems with using objects.

This has now been fixed so the bool is set every `OnTriggerStay`
if no object is being touched or if the object being touched is the
same as the current touched object.