feat(smoothing): add generic smoothing and pointer smoothing settings
Most humans aren't able to hold controllers completely still. To
improve the user experience it is therefore a good idea to smooth the
position and rotation of tracked objects like controllers.

This is enabled by improving on the existing `VRTK_ObjectFollow`. This
change adds two new subclasses to support smoothing any combination of
position, rotation and scale of a game object's transform or rigidbody.

One of the best candidates for smoothing are the various pointers in
VRTK. That's why this change also adds smoothing settings to the base
pointer so both the simple and bezier pointer benefit of a smoothing
option.

Thanks to Nathan Rowe from SculptrVR for the pseudocode!