fix(SDK): run controller init code using Awake
The `Awake()` method is supported in the ScriptableObject so it's
more appropriate to use this and `[RuntimeInitializeOnLoadMethod]`
only works on static methods meaning the `Initialise()` method was
never being run.

The Oculus SDK doesn't need an Awake method as the controllers
are not ready to be cached at that time anyway and the cache setup
will be run the first time they are called, which is a more
appropriate time to set the cache up as the other dependencies will
be available.