fix(SDK): ensure SteamVR returns false for button two touch
The SteamVR SDK has no mechanism for a button two on a controller but
the SDK for the touch event was still returning the value as if the
Application Menu button on the controller had been pressed. It should
return false like the other SteamVR button two SDK calls do.