fix(SDK): ensure consistent button mapping on oculus touch for steamvr
  > **Breaking Changes**

Oculus Touch through SteamVR exposes the additional touch buttons but
also maps the Vive application menu button through to `Button Two` on
the Touch controller. This means that the Menu button in the
Controller Events script now should be set to `Button Two` and SteamVR
SDK now treats `Button Two` as the application menu button.

This will break things for some users if they have Button One mapped
as the menu button, because on button one isn't associated to any
button on the Vive wand.

The Oculus Touch also has a different controller model through SteamVR
and therefore the SteamVR Controller SDK script now listens for the
connected headset and provides different paths and colliders for the
controller based on which one is connected.

The Oculus Touch models are also at a rotated orientation even though
their forward matches the Vive wand, which means things like the
pointers are all out of alignment. To solve this, a custom transform
is returned for the pointer origin if the Rift is being used with
SteamVR.