fix(Interaction): ensure auto grab works correctly with prefabs
If a Prefab was attempted to be auto grabbed that was also set to
have `Disable On Idle` as true, then the prefab would clone and be
disabled at the point of auto grabbing and therefore drop to the
ground.

This fix adds an additional option to specify the `Object To Grab`
is a prefab and if it is then to disable the `Disable On Idle` check
during the auto grab and to reset it when it has completed.

Another option has been added that determines whether to always clone
the object when the script is re-enabled or whether to just attempt
to grab the original clone.