fix(SnapDropZone): save interactable object state before force snap
Upon force snapping to a snap drop zone, the interactable object state
needs to be set, because if it is snapped then the SaveCurrentState
method won't be run and the previous parent will not be saved causing
it to lose it's parent when it is unsnapped and dropped.

Calling SaveCurrentState before the force snap takes place means the
previous parent will be adhered to when it is dropped.