feat(SDK): allow for dynamic SDK discovery
The list of supported SDKs in VRTK is fixed because they are hard coded
into the SDK manager (and its editor). This change modifies the SDK
manager so it looks up the available SDKs by checking the code base for
classes that inherit from one of the base SDK classes. This allows to
create and use additional SDKs without touching the existing VRTK
codebase at all.

Sharing of SDK implementations in a private group of people (e.g. where
everyone signed an NDA) is now possible without needing to share patches
to the VRTK codebase. There is no reason anymore to keep up with changes
to the parts of VRTK one changed to use their own SDK implementations.

To allow for easy specifying of SDK implementations a new "SDK
description" attribute is added. Another "predicate" attribute allows to
mark a static method as a predicate for SDK purposes. The method signals
whether to add or remove the scripting define symbol that is specified
by the attribute by returning true or false respectively. Marking the
same method with multiple predicate attributes is allowed.

The existing SDK implementations have been updated to make use of both
attributes. Usage of the SteamVR plugin in VRTK is now simplified by
using the predicate attribute to add a scripting define symbol that
defines what version of the plugin is currently used in addition to
defining the generic SteamVR SDK symbol.

If a set SDK doesn't exist anymore (e.g. the SDK class was removed) the
SDK manager will use the corresponding fallback SDK instead and log an
error so the problem is known about. The editor shows that error, too,
but won't remove the set SDK from its list of selectable SDKs until the
user overrides it with another SDK. It shows the name of the missing
class in that case instead.

The SDK manager's editor is updated to allow selecting one of the
discovered SDKs. Logic that previously handled the scripting define
symbols is now moved to the SDK manager class itself in form of public
methods. This allows to build scripts that handle automatic building for
different stores as Oculus Home doesn't allow to include SteamVR/OpenVR
libraries, for example. The automatic populating of the SDK object
references like the Linked Objects and Alias Objects is moved from the
editor to a new setting on the SDK manager itself. This has a public
method to trigger manually, too.

Additionally the editor was changed to support undo/redo for all the
possible changes. Proper error messages and hints are now shown in the
editor if needed, too.

The existing auto-managing of the scripting define symbols is now done
whenever needed. This means loading a scene with an active SDK manager
in it will make sure to manage the symbols if the SDK managers setting
says to do so. The behavior for the auto-populating of the SDK object
references is the same with this change. The idea is to always make sure
the current setup is correct for a used SDK manager.