fix(Interaction): prevent rotator track applying velocity on release
The rotator track mechanic should not apply thrown velocity when the
interactable object is released otherwise it causes the joint to
bounce back.

The fix for this is to override the `StopGrab` method on the Track
Object and prevent it from doing the ReleaseObject method which
is what adds the velocity.