feat(Pointer): allow overriding of base pointer controller events
The pointer events by default all link to the AliasPointer events,
however it may be required to override those events with a custom
button press. It's now possible to override the button press and
have two pointers with different activation buttons by extending
a pointer script and overriding the `AliasRegistration` method.

In the future, there may be an inspector option to provide a button
override instead of needing to do it via script.