fix(Interaction): provide more natural door control
  > **Breaking Changes**

The 3D Door Control now uses the RotatorTrack grab mechanic to provide
a more natural door operation. The `Snapping` inspector parameter has
now been replaced with a float to determine the snap distance the door
has to be from being in a closed state before it snaps shut.

It is also possible now to provide the friction value that the door has
when it is being grabbed and moved, and the friction when it is not
being grabbed. A high grabbed friction gives a less bouncey feel and a
low ungrabbed friction allows the door to swing freely to allow for
slamming shut.

Another tweak is it is possible to apply the interactable object to the
entire door object rather than just the handle, which can be useful
for small doors such as cupboards, where it's desired to grab the door
to close it.