feat(Interaction): add wheel 3d control
The Wheel 3D control provides a spinnable object that represents a
wheel. It can be turned by grabbing the object like a steering wheel
or it can be used like a dial where a rotational force is used to
spin the wheel.

It is possible to provide an angle limit on the wheel to prevent it
from being turned beyond a max and min point. It is also possible
to have the wheel snap to the absolute angle for the current number
value making it useful for dials.

The wheel is very similar to the 3D knob control and may eventually
supercede the knob control in the future.

The example scene for 3D controls has been updated to contain a
selection of wheels and dials to demonstrate the new functionality.

A teleporter has also been added to the example scene to make it
easier to reach the relevant controls for testing.