fix(Interaction): rebuild slider control to use joints
  > **Breaking Changes**

The old Slider 3D Control used world positions to determine the slider
placement, this caused issues when the slider was attached to something
and it also meant it could not be attached to something that moved
around.

Another fundamental issue occured when the slider was placed along a
world z-axis as the world positions would cause the slider to
constantly flicker between the two extreme values.

The solution has been to re-write the slider and use a configurable
joint to allow the slider to move across the axis. It is also required
now that the min/max points of the slider are provided as a collider
which also keeps the slider from exceeding the values using game
physic collisions.

The slider snap is also resolved with physics now using the spring
position on the joint.

It is also now possible to attach the slider to a moving object and
it will still function.