feat(Interaction): add friction/spring options to lever/spring lever
The Lever and Spring Lever 3D controls now both have the option to
set the lever angular drag when the lever is being grabbed and when it
has been released.

The Spring Lever now also has options to set the spring parameters
and the spring can be disabled when grabbing.