fix(Locomotion): prevent touchpad control of play area when falling
Previously, it was possible to continue to control the play area when
falling. However, this causes issues as controlling the play area
whilst falling is unnatural.

This fix utilises the `IsFalling()` method on the Body Physics script
to determine if the player is falling and prevents the touchpad from
controlling the PlayArea until the player is no longer falling.

If no Body Physics script is present then the touchpad can still
control the PlayArea during a fall.

The LeftController/RightController getter/setters have been replaced
and the private variables have just been made public as toggling
the variables at runtime in the editor was not calling the setters.

As the alternative logic is so simple, it's worth just doing it that
way for the benefit of inspector control.