fix(Locomotion): prevent MoveInPlace control work whilst falling
Previously, it was possible to continue to control the play area when
falling. However, this causes issues as controlling the play area
whilst falling is unnatural.

This fix utilises the `IsFalling()` method on the Body Physics script
to determine if the player is falling and prevents the player from
controlling the PlayArea until the player is no longer falling.