fix(Climbing): add moving and rotating to VRTK_PlayerClimb
Climbable move points will now move and rotate the player as they
do likewise around the game world. You can disable the rotation of
the player (for comfort mode) by the inspector toggle named
useGrabbedObjectRotation in the VRTK_PlayerClimb class. With it off
the player will only pivot around the grabbed location, but will
not rotate the view at all.