feat(Locomotion): add snap rotation touchpad control action
A new touchpad control action has been added that causes the controlled
object to snap rotate around it's `up` vector based on the position of
the touchpad axis.

This provides a sudden change in rotation angle as opposed to the
smooth angle change of the Rotate touchpad control action.

It's also possible to provide a delay between each rotation and an
optional fade/blink between each rotation to reduce nausia.

This can be useful for rotating the play area for seated games.