feat(Locomotion): use events with touchpad control instead of scripts
The Touchpad Control script has now been updated to use events to let
the Control Actions know when to perform their action rather than have
the Touchpad Control script need to know about what action is going to
happen.

This means that the Touchpad Control script is agnostic of the action
that happens upon an axis changing and it also means that multiple
actions can be registered to happen upon the touchpad axis changing.

The example scene has been updated with a collection of game objects
that contain different variations of control actions that can be
toggled on and off to demonstrate different configurations.