fix(Pointer): use capsule collider for pointer tip interaction
There is an issue when using a sphere collider that it has to be a
substantial size larger than the pointer tip for it to continually
register collisions when the pointer is being moved fast.

The CapsuleCollider provides a better shape that can be projected
forward to get better penertration and keep the collider active when
the tip is close by.

The creation of these objects has also been separated out into smaller
protected methods so if required they can be overriden and replaced
with an alternative collider.