feat(Interaction): Handed events/actions
This changeset adds the following members to `ControllerEvent` and
`ControllerAction` classes:
-  `public SDK_BaseController.ControllerHand trackedHand` field
-  `public GameObject ControllerAlias` property.

The new `trackedHand` enum property specifies which of the controllers a
`ControllerEvents` instance should emit events for. This shows up as a
handy drop-down in the editor.

If `trackedHand` is `None` then `ControllerAlias` will resolve to the
local `GameObject` of the `ControllerEvents` instance. Otherwise, the
property will use the `trackedHand` field to look up the appropriate
"Script Alias" object on the `SDKManager` singleton; either
`scriptAliasLeftController` or `scriptAliasRightController`.