fix(Pointer): update interactor collider in FixedUpdate
The interactor collider is updated to match the current pointer
position. Previously, it was done in the Update method but this meant
that the interactor collider would lag behind the pointer tip causing
issues.

Putting the update of the interactor collider in the FixedUpdate
method means it is updated more frequently and therefore it has a
higher level of fidelity removing any glitches that were previously
showing up.