fix(Interaction): only set force dropped when stop grabbing
Previously, the `forceDropped` was set to true when stop touching and
stop using was called. This would cause issues when an object was
disabled and re-enabled as the previous state would be loaded even
though it hasn't currently been saved.

There is no reason to set `forceDropped` to true on stop touch or
stop use as it's only relevant when an object is being grabbed because
this is the only way it can be force dropped.