fix(ControllerEvents): ensure axis events take place on small changes
Previously, the touchpad axis would not update if the touchpad was
touched slowly creating a small delta because the previous touchpad
axis state was always being updated even if a change hadn't occured.

The fix is to ensure that the previous touchpad axis state is only
updated when the touchpad axis has actually changed. This allows for
smaller movements to be registered still.

The Touchpad Axis Change event is also only emitted now when the
touchpad is being touched as it would previously emit when the touchpad
was released causing an additonal axis event change that wasn't
required. The Trigger and Grip axis change events also have their axis
value forced to `0f` now if the axis is no longer being touched as
sometimes the controller hardware could report an untouch of a button
but the axis was still reporting a value over 0f which seems wrong.