fix(Pointer): ensure visibility and interaction states are toggled
There was an issue where the cursor and tracer visibility could not
be toggled to always visible unless a toggle event was called.

There was another issue where if the tracer or cursor was active
that the interaction object would not be enabled until the pointer
activation button was pressed.

This has now been fixed by ensuring the correct state of the tracer
and cursor visibility is checked taking into account the visibility
state. Also, if the tracer or cursor is visible even when the
pointer activation button hasn't been pressed then the object
interactor will be activate.