fix(Locomotion): force controller listener setup on destination points
The destination point script now forces the controller listener
registration in the OnEnable to ensure that if the controller is not
on prior to running the scene then the listeners are still registered.

This fixes an issue where the destination points would not work if
the controllers were off at the point of starting the scene.