fix(Pointer): prevent flickering when activating an always on pointer
If the tracer or cursor was set to always on and the activation button
was pressed then the cursor would be set to disable renderers and
re-enable next frame which would cause the flickering.

This has now been resolved by only calling the disable if the pointer
isn't using always on.

Another issue has been fixed with the play area cursor not coming on
and staying on at the right times. It should always come on when the
cursor is active as it's an extension of the cursor.

The documentation has also been generated, which includes the docs
for the pointer smoothing work as well.