fix(Highlighters): ensure custom outline arrays are never null
The `customOutlineModels` and `customOutlineModelPaths` arrays could
be null if they were not set in the inspector and the scene changes
to a scene that uses a highlighter because Unity only seems to
auto initialise the array if the scene the array is being used in is
the one that is started.

This fix ensures to set the array if it's null.