fix(Controls): ensure door force doesn't get incorrectly applied
There were issues with how the door snap shut force was being applied
to doors of a certain rotation due to the Third Direction variable
not providing an accurate way of determining the door open direction.

This has been fixed by using the limits on the hinge joint to determine
the direction the door opens.