feat(Pointer): add selection on hover parameter
It is now possible to initiate the pointer select action based on if
the pointer is hovering over the same collider for a specified length
of time.

Another parameter has been added that activates the pointer beam when
the Pointer script is enabled.

This now allows for pointer beams to be used to make selections without
the need of any controller input button being pressed.

This means it's possible to use pointers on the headset without a
controller requirement, which is useful for Headset only scenes.

The error message that was thrown when a controller was not given has
now also been removed as a controller is no longer essential.