feat(Interaction): add lower threshold for controller axis events
The controller axis for trigger and grip now both have a lower
threshold which can be set to determine when the axis should be at 0f
as the controller hardware seems to malfunction and report various
axis numbers when the button is not being touched.

Another parameter has been added to force the axis to 0f when the
untouch event of the button happens as the axis shouldn't report
anything other than 0f if the button isn't being touched in some
occasions.