fix(Presence): set crouch rewind threshold to stop fall through floor
There was an issue where the Position Rewind script was causing the
play area to fall through the floor if the scene was started with
the headset on the floor.

This has been fixed by preventing the setting of the initial position
if the headset is lower than the required crouch position.

An additional parameter has been added that will prevent rewind from
taking place at all if the headset is below a certain Y threshold,
which can prevent the headset causing rewinds if the user places
the headset on the floor.