fix(Climbing): scaled throw from climb
When the play area is scaled, throwing yourself from a
hand hold resulted in a double scaled velocity. It was
being applied in VRTK_PlayerClimb as well as when transforming
the vector to world space in BodyPhysics. This fix removes the
scale from happening in BodyPhysics, thus keeping the input
parameter as a world space input, and changing the value
instead in PlayerClimb (to world space orientation) before
passing the velocity along.