fix(EventSystem): screen space overlay desktop camera (again)
There's an issue where a "Screen Space - Overlay" camera doesn't
register any mouse input. The problem (this time) is that
`VRTK_EventSystem.SetEventSystemOfBaseInputModulesTo` is called too
early and therefore never changes the event system reference for some of
the existing input modules.

The fix is to run `VRTK_EventSystem.SetEventSystemOfBaseInputModulesTo`
after a frame has passed by making it a coroutine.