fix(SDKManager): enable auto populate button at all times
The inspector for the SDK Manager doesn't allow clicking the auto
populate button if the manager is set up to populate the object
references automatically. Since the auto-populating only occurs whenever
an SDK selection is changed this lead to a problem when changing any of
the objects to reference.

To allow populating the object references manually this fix makes sure
to enable the button at all times.