fix(SDKManager): handle scene reload
The SDK Manager made sure to not trigger the managing of the scripting
define symbols and population of the object reference when a scene is
reloaded. This behavior is wrong when the following is done:
- Change which SDKs are used or which objects are referenced in the SDK
Manager.
- Wait for it to change the scripting define symbols (if needed).
- Reload the scene **without** saving the changes done to the SDK
Manager.
This results in a scene reload that neither reverts the scripting define
symbols, nor the object references according to the previous
configuration of the SDK Manager.

This fix makes sure to always manage the scripting define symbols and
populate the object references on a scene reload.