fix(Presence): use lower precision on checking y position for falling
There was an issue where the play area would report it is still falling
even though the fall looked as if it had stopped. This was due to the
check between the current play area Y position and the previous Y
position being very slightly different. This has been fixed by adding
a new parameter that allows the setting of the number of decimal places
to round the Y positions to when doing the comparison.

5 decimal places seems to be the most appropriate number for general
uses. The round float method has been added to the SharedMethods as
it may be useful for other things in the future.