feat(Simulator): add mouse input mode that requires a button press
The default simulator behaviour of always accepting mouse input can be
problematic because mouse input terminates at the edge of the monitor
and can leave you in situations where you can't easily rotate/move as
far as you need. Or your mouse may be over the editor instead of the
game window when you need to grab something with a mouse click.

This implements an optional alternative mouse input method that works
similarly to the scene view where mouse movement only rotates/moves
when a button (such as the right mouse button) is pressed.