feat(Haptics): extract haptics out of ControllerActions
The haptic feedback has now been removed from the ControllerActions
script and placed in the Shared Methods script meaning that it can
be called simply by knowing the ID of the controller to rumble.

This means it no longer requires a script to be attached to a script
alias object to achieve haptic feedback.

A new Instance Methods script has been added that handles running
any co-routines such as the haptic feedback co-routine in a global
manner rather than being tied to a specific controller objeect.