fix(Locomotion): prevent unexpected change in direction in MoveInPlace
The MoveInPlace script would change direction in controller direction
mode if the controllers were inverted during running. This could happen
by simply running too hard and the controller would go past a certain
angle meaning the direction would change causing a unwanted jerky
motion.

This has been fixed by keeping a record of the previous direction
and if the angle between the two directions is too great then it means
the controllers have become inverted and the previous direction is
used for the forward direction.