fix(UICanvas): ghost in layout group
The draggable helper game object the UI Canvas uses interferes with the
layout system and shows up as an empty "ghost" object when used in
conjunction with a layout group. The fix is to add a `LayoutElement` and
setting `ignoreLayout` to true so the layout system ignores that element
completely.

This fix resolves #925.