fix(Locomotion): ensure current destination point is reset to null
The current destination point was never reset to null if a destination
point was not being used. This meant that destination points could
not be used along side normal teleporting as the current destination
point was not being set to null so even if the teleport location was
somewhere other than a destination point, the last destination point
would still be considered current and therefore wouldn't reactivate.