fix(Internal): ensure tracked controller index is set correctly
The Tracked Controller script index parameter was being set too
aggressively meaning if the SDK didn't actually have an index to
provide then the index would be reset on the next frame to MaxValue
causing OculusSDK to not work.

The solution is to just remove the use of the `currentIndex` private
variable and just update and use the `index` public variable as this
will ensure there is only 1 true index variable being used and it is
updated at the correct time.