fix(OculusAvatar): define symbol
The scripting define symbol for the Oculus Avatar SDK was never added.
This meant the Oculus SDK didn't work in VRTK if the Avatar SDK was
used, too.

The fix is to make sure to add scripting define symbols even if they
don't are the symbols specified in the currently selected `SDKInfo`s.
Additionally the correct case for the Avatar SDK class is used to make
sure the type is actually found if the SDK is installed.