fix(UICanvas): RectTransform in 5.5.2p2 and 5.6
Unity changed the way additions of components derived from `Transform`
are handled:
- From the 5.5.2p2 changelog:
>Kernel: Prevent scripts from adding Transform-derived components during
their Awake methods.

- From the 5.6.0b11 changelog:
>UI: Adding UI elements via script will cause OnDisable and OnEnable to
be called on all scripts in the hierarchy if a RectTransform needs to be
added.

To fix the first problem the creation of any game object that uses
`RectTransform` is deferred to the next frame. The second problem is
fixed by specifying the `RectTransform` to add in the game object
constructor.