feat(SDKBridge): cache invalidation
The SDK Bridge caches the SDKs returned by the SDK Manager instance if
it exists. To allow runtime switching of SDKs it is needed to be able to
invalidate those cached objects. This change adds a new method to the
SDK Bridge to allow doing just that and modifies the SDK Manager to call
said method whenever the SDK Manager instance changes.