fix(SDKs): allow setup at runtime
The SDKs check for existing objects in the scene by using
`GameObject.Find` and other methods. These methods only return objects
which are in the scene and active. To allow runtime setup of the SDK
manager it is important to be able to add the needed objects for the
SDKs before or after setting the SDK Manager to use said SDKs. This fix
makes sure setting up the SDK Manager is possible at runtime by adding
and using a new method that allows to also find inactive objects in a
scene.