feat(Simulator): lock mouse by default and hints background
The Simulator allows to lock the mouse when controlling the hands. Since
this is very useful in the Editor it is now enabled by default.
Additionally the Simulator Camera Rig prefab has been updated to make
sure the hints text is always in the top left corner of the screen and
has a background to make reading the text easier.