fix(Interaction): allow grab on force snapped item after ungrab
There was an issue with interactable objects that were set to not
swappable if they were force snapped to a SnapDropZone and then grabbed
the previous grabbable state would be false by default. This meant that
if the object was grabbed out of the SnapDropZone and ungrabbed then
the `isGrabbable` bool would be false meaning the object could not be
grabbed.

The fix was to ensure the `previousIsGrabbable` bool is set in the
`SaveCurrentState` method.