fix(Interaction): force secondary controller release on primary drop
There was an issue where the secondary controller was not notified of
the object release if the secondary controller was grabbing whilst the
primary controller was released.

The action is that the object is dropped and the secondary controller
has no influence over the object any more, but the secondary
controller Interact Grab script was not notified of the drop so it
would not emit the event and would keep the state of a grabbed object
until the grab button on the secondary controller was released.

This has been fixed by calling `ForceRelease` on the secondary
controller if the primary controller is ungrabbed. As the secondary
controller grab actions are cleaned up before the primary controller
then a `ForceRelease` should simply clean up the state of the
secondary controller Interact Grab script and not actually force
drop the object if it's still being held by the primary controller.