feat(Utilities): add custom raycast script for customising rays
Many of the VRTK scripts utilise raycasts but have the raycast options
hard coded in the script. This new Custom Raycast script allows any
VRTK script that uses a raycast to have the raycast customised by
injecting a Custom Raycast into it.

Previously, the VRTK scripts that utilised a raycast would have a
`Layers To Ignore` parameter which would be passed through to the
raycast. However, this has now been deprecated as all of these scripts
can achieve the same customisation by using a Custom Raycast instead.