feat(sdk): implement support for gear vr
This is the PR #944 by @CoderDawson updated to the latest VRTK version,
modified to be compliant with the new SDK management and with updated
documentation.

Add SDK_GearVRDefines to handle the scripting symbols for the GearVR.

Refactor and update other GearVR/SDK files.

Update documentation (`GearVR_README.md`).

Tested with GearVR:
* object touching and grabbing (005_Controller_BasicObjectGrabbing)
* interaction with Unity UI (034_Controls_InteractingWithUnityUI)
* teleport (044_CameraRig_RestrictedTeleportZones)

The examples were tested after changing `[CameraRig]` and `[VRTK]` in
order to support GearVR.