fix(SnapDropZone): allow for child colliders to activate drop zone
The Snap Drop Zone policy list check was always done on the game object
for the collider touching the drop zone, however this would mean that
if a collider was a child of the interactable object then the policy
list would disallow the drop as it needed to check the validity of the
interactable object's game object.

This has now been fixed to check the interactable object's game obejct
in the policy list check.