feat(Pointer): allow custom valid/invalid play area cursor object
The Play Area Cursor can now accept a valid and an invalid prefab
parameter to customise the look of the play area cursor when it is
not colliding and when it is colliding with a valid object.

It's also possible to disable the pointer colour inheritance if a
more customised cursor is required.

Also, the custom play area cursor now resizes to the actual play area
boundaries sizes unless a custom size is entered.