fix(SDK): caching SDK Manager objects methods
The SDK bases Boundaries and Headset provided methods to their
subclasses that made it easier to get the SDK Manager's play area and
headset, respectively. The methods themselves and their usage didn't
result in the intended caching of the returned objects. Since the
subclasses do the caching properly already the methods and their usage
is removed in this fix.