fix(UpdatePrompt): prevent null path to version file
When checking for an VRTK update the Update Prompt tries to find the
version file by looking at the directory the script is in. This fix
makes sure to throw an exception on not finding the script instead of
using a `null` path, therefore making the issue more obvious in case
there is an issue finding the version file at all.