fix(ControllerEvents): prevent events emitting if object not active
There was a rare condition in the Object Appearance script where it
would attempt to emit controller events on an inactive controller
game object which would cause a crash when changing scenes.

This has been fixed by only emitting the event if the controller
object is active in the scene.