fix(Interaction): update interact touch parameter to make sense
  > **Breaking Changes**

The `customRigidbodyObject` parameter in the Interact Touch script has
been changed to `customColliderContainer` as that's actually what the
parameter is for as it has nothing to do with rigidbodies and only to
do with the compound colliders used for touching.

Whilst, this causes a breaking change, it's worth doing as it's just
simply misleading.