fix(Locomotion): ensure play area collider is found in children
The Base Object Control Action script requires the collider on the
Playarea that is added by the Body Physics script. The Body Physics
script no longer adds the collider to the root Playarea object and
it is instead added as a child object.

This fix ensures children are also looked in for the collider.