fix(Interactions): leave usage when disabling InteractUse
  > **Breaking Changes**

Whenever the InteractUse script was disabled it stopped the usage of the
currently held object by the controller but also stopped the object's
using action. This fix makes sure to only force the controller to stop
the usage, not the object itself.

If users of VRTK depend on the old behavior they should stop the
object's usage themselves when the stopping is recognized using one of
the events or virtual methods.