fix(Presence): ensure is moving isn't always toggled on in body phsyics
The Body Physics script was continually toggling the isMoving parameter
even when standing still due to the ray cast collection that checks to
see if the user is walking over an uneven floor.

This was because the system casts 2 rays down, 1 from the headset
position and 1 just outside of the body collider and checks to see if
the difference means an incline/decline has occured. The issue was
that the headset rotation was affecting the forward direction of the
second down cast.

Another additional parameter has been added to allow the play area
movement threshold to be determined to prevent the headset pushing
the play area back if the headset is moved very slowly.

The code has also been cleaned up to make it easier to read.