fix(Interaction): stop disable code running if object starts disabled
If the Interactable Object is set as Disable When Idle, then upon first
awake it will enable the immediately disable itself and call the
OnDisable code, which isn't desired or required as it means the object
will call any disable code when it's supposed to start in a disabled
state.

The solution is to add a quick check to see if the object is supposed
to start in it's disabled state and then not do any of the relevant
disable code when the OnDisable is called.

Cleaning up anything created in the OnEnable is still done.