fix(SDK): ensure simulator returns valid play area vertices array
The Simulator was returning a vertices array that was too small and
therefore caused anything that required those vertices to crash.

Also, all SDKs need to return null if a valid vertices array cannot
be obtained so the consuming script can do a null check and prevent
attempting to build something with an invalid vertice array.