fix(SDK): ensure SteamVR works by default in Unity 5.6
There is an issue with SteamVR 1.2.1 and below and Unity 5.6
where the VR camera does not have the SteamVR_UpdatePoses script on
it and therefore the controllers do not appear.

This check simply adds the required script if it is missing. This fix
will probably be removed in the future when SteamVR is fixed.