fix(Interaction): ensure joint break force is set correct on drop zones
When an Interactable Object is set to
`Valid Drop = Drop Valid Snap Drop Zone` it ensures the `IsDroppable`
method returns false if the object is not in a snap drop zone then
on grabbing with a joint based grab mechanic it sets the break force
on the joint to `infinity`. However, if the object is grabbed from
a snap drop zone, then the `IsDroppable` will be true and therefore
the joint break force wasn't set to `infinity` meaning an object
could be dropped by breaking the joint.

This fix ensures that can't happen by doing an additional check when
setting the joint break force.