fix(SnapDropZone): prevent enter zone event firing multiple times
The `OnObjectEnteredSnapDropZone` event would fire multiple times when
an object was held over a drop zone collider.

This has been fixed now by doing a simple check if the same object is
being held in the drop zone and only fire the event on the first
occurance.

It will also fire the enter event when the object is removed from the
drop zone as it has been unsnapped and is in the drop zone collider
space again.