fix(SnapDropZone): allow multiple valid objects to hover at same time
The Snap Drop Zone can now have multiple valid snappable objects in
the hover zone and any object that is released in the snap drop zone
will be snapped.

This fixes an issue where previously only the first valid object
that entered the snap drop zone hover space would be able to be
snapped and any other object even if it was valid would not be snapped
and would just fall on drop.