feat(DestinationMarker): add hover event
The destination marker now also has a hover event state that can
be used to determine if the destination marker is still hovering over
the existing target collider. This enables the Enter event to only
be emitted when the destination marker first enters the collider.

This provides a cleaner solution as Enter/Exit can now be called
only once whilst hover can be called multiple times if required.

The VRTK_Pointer has been updated to call hover if the collider
hasn't changed, and to call enter if the collider does change.

As the Enter event is no longer emitted every frame, this may cause
issues with solutions that were listening for the Enter event for
changes in the destination marker position.

This can still be achieved by calling
`SetForceHoverOnRepeatedEnter(false)` on the destination marker
which will not do the additional check when calling the Enter event
that propagates the continued enter to the Hover event.